Former-commit-id: a7afb1de4f47db2668f68bd8fa937dfea897eb80

+++ source/Environments/Environment.h
@@ -31,7 +31,7 @@ protected:
     void commonSetup();
     void spawn_server();
 public:
-    int iter;
+    long unsigned int iter;
     bool resetAll;
     vector<Agent*> agents;
     vector<Real> max_scale, min_scale;
